+++ examples/03_datetime_encoder.py
@@ -35,8 +35,8 @@ It is used by default in the |TableVectorizer|.
 .. |HGBR| replace::
     :class:`~sklearn.ensemble.HistGradientBoostingRegressor`
 
-.. |to_datetime| replace::
-    :func:`~skrub.to_datetime`
+.. |ToDatetime| replace::
+    :class:`~skrub.ToDatetime`
 """
 
 
@@ -44,63 +44,54 @@ It is used by default in the |TableVectorizer|.
 # A problem with relevant datetime features
 # -----------------------------------------
 #
-# We will use a dataset of air quality measurements in different cities.
-# In this setting, we want to predict the NO2 air concentration, based
-# on the location, date and time of measurement.
+# We will use a dataset of bike sharing demand in 2011 and 2012.
+# In this setting, we want to predict the number of bike rentals, based
+# on the date, time and weather conditions.
 
 from pprint import pprint
 
 import pandas as pd
 
 data = pd.read_csv(
-    "https://raw.githubusercontent.com/pandas-dev/pandas"
-    "/main/doc/data/air_quality_no2_long.csv"
-).sort_values("date.utc")
+    "https://raw.githubusercontent.com/skrub-data/datasets/master"
+    "/data/bike-sharing-dataset.csv"
+)
 # Extract our input data (X) and the target column (y)
-y = data["value"]
-X = data[["city", "date.utc"]]
+y = data["cnt"]
+X = data[["date", "holiday", "temp", "hum", "windspeed", "weathersit"]]
 
 X
 
 ###############################################################################
-# We convert the dataframe date columns using |to_datetime|. Notice how
-# we don't need to specify the columns to convert.
-from skrub import to_datetime
+# We convert the dataframe ``"date"`` column using |ToDatetime|.
 
-X = to_datetime(X)
+from skrub import ToDatetime
+
+X["date"] = ToDatetime().fit_transform(X["date"])
+
 X.dtypes
 
 ###############################################################################
 # Encoding the features
 # .....................
 #
-# We will construct a |ColumnTransformer| in which we will encode
-# the city names with a |OneHotEncoder|, and the date
-# with a |DatetimeEncoder|.
+# We directly encode the ``"date"`` column with a |DatetimeEncoder|.
 #
 # During the instantiation of the |DatetimeEncoder|, we specify that we want
 # to extract the day of the week, and that we don't want to extract anything
-# finer than minutes. This is because we don't want to extract seconds and
-# lower units, as they are probably unimportant.
-
-from sklearn.compose import make_column_transformer
-from sklearn.preprocessing import OneHotEncoder
+# finer than hours. This is because we don't want to extract minutes, seconds
+# and lower units, as they are unimportant.
 
 from skrub import DatetimeEncoder
 
-encoder = make_column_transformer(
-    (OneHotEncoder(handle_unknown="ignore"), ["city"]),
-    (DatetimeEncoder(add_weekday=True, resolution="minute"), "date.utc"),
-    remainder="drop",
-)
+date_enc = DatetimeEncoder().fit_transform(X["date"])
 
-X_enc = encoder.fit_transform(X)
-# pprint(encoder.get_feature_names_out())
+print(X.date, "\n\nHas been encoded as:\n\n", date_enc)
 
 ###############################################################################
-# We see that the encoder is working as expected: the ``"date.utc"`` column has
-# been replaced by features extracting the month, day, hour, minute, day of the
-# week and total second since Epoch information.
+# We see that the encoder is working as expected: the ``"date"`` column has
+# been replaced by features extracting the month, day, hour, day of the
+# week and total seconds since Epoch information.
 
 ###############################################################################
 # One-liner with the |TableVectorizer|
@@ -116,12 +107,13 @@ pprint(table_vec.get_feature_names_out())
 
 ###############################################################################
 # If we want to customize the |DatetimeEncoder| inside the |TableVectorizer|,
-# we can replace its default parameter with a new, custom instance:
+# we can replace its default parameter with a new, custom instance.
 #
-# Here, for example, we want it to extract the day of the week.
+# Here, for example, we want it to extract the day of the week:
 
-table_vec = TableVectorizer(datetime=DatetimeEncoder(add_weekday=True)).fit(X)
-pprint(table_vec.get_feature_names_out())
+# use the `datetime` argument to customize how datetimes are handled
+table_vec_wd = TableVectorizer(datetime=DatetimeEncoder(add_weekday=True)).fit(X)
+pprint(table_vec_wd.get_feature_names_out())
 
 ###############################################################################
 # .. note:
@@ -130,7 +122,7 @@ pprint(table_vec.get_feature_names_out())
 #
 # Inspecting the |TableVectorizer| further, we can check that the
 # |DatetimeEncoder| is used on the correct column(s).
-pprint(table_vec.transformers_)
+pprint(table_vec_wd.transformers_)
 
 ###############################################################################
 # Prediction with datetime features
@@ -139,18 +131,13 @@ pprint(table_vec.transformers_)
 # For prediction tasks, we recommend using the |TableVectorizer| inside a
 # pipeline, combined with a model that can use the features extracted by the
 # |DatetimeEncoder|.
-# Here's we'll use a |HGBR| as our learner.
+# Here we'll use a |HGBR| as our learner.
 #
-# .. note:
-#    You might need to require the experimental feature for scikit-learn
-#    versions earlier than 1.0 with:
-#    ```py
-#    from sklearn.experimental import enable_hist_gradient_boosting
-#    ```
 from sklearn.ensemble import HistGradientBoostingRegressor
 from sklearn.pipeline import make_pipeline
 
 pipeline = make_pipeline(table_vec, HistGradientBoostingRegressor())
+pipeline_wd = make_pipeline(table_vec_wd, HistGradientBoostingRegressor())
 
 ###############################################################################
 # Evaluating the model
@@ -176,70 +163,49 @@ cross_val_score(
 # Plotting the prediction
 # .......................
 #
-# The mean squared error is not obvious to interpret, so we compare
-# visually the prediction of our model with the actual values.
+# The mean squared error is not obvious to interpret, so we visually
+# compare the prediction of our model with the actual values.
+# To do so, we will divide our dataset into a train and a test set:
+# we use 2011 data to predict what happened in 2012.
 import matplotlib.pyplot as plt
 
-mask_train = X["date.utc"] < "2019-06-01"
+mask_train = X["date"] < "2012-01-01"
 X_train, X_test = X.loc[mask_train], X.loc[~mask_train]
 y_train, y_test = y.loc[mask_train], y.loc[~mask_train]
 
 pipeline.fit(X_train, y_train)
 y_pred = pipeline.predict(X_test)
 
-all_cities = X_test["city"].unique()
+pipeline_wd.fit(X_train, y_train)
+y_pred_wd = pipeline_wd.predict(X_test)
 
-fig, axes = plt.subplots(nrows=len(all_cities), ncols=1, figsize=(12, 9))
-for ax, city in zip(axes, all_cities):
-    mask_prediction = X_test["city"] == city
-    date_prediction = X_test.loc[mask_prediction]["date.utc"]
-    y_prediction = y_pred[mask_prediction]
-
-    mask_reference = X["city"] == city
-    date_reference = X.loc[mask_reference]["date.utc"]
-    y_reference = y[mask_reference]
-
-    ax.plot(date_reference, y_reference, label="Actual")
-    ax.plot(date_prediction, y_prediction, label="Predicted")
-
-    ax.set(
-        ylabel="NO2",
-        title=city,
-    )
-    ax.legend()
-
-fig.subplots_adjust(hspace=0.5)
-plt.show()
-
-###############################################################################
-# Let's zoom on a few days:
-
-mask_zoom_reference = (X["date.utc"] >= "2019-06-01") & (X["date.utc"] < "2019-06-04")
-mask_zoom_prediction = (X_test["date.utc"] >= "2019-06-01") & (
-    X_test["date.utc"] < "2019-06-04"
+fig, ax = plt.subplots(figsize=(12, 3))
+fig.suptitle("Predictions with tree models")
+ax.plot(
+    X.tail(96)["date"],
+    y.tail(96).values,
+    "x-",
+    alpha=0.2,
+    label="Actual demand",
+    color="black",
+)
+ax.plot(
+    X_test.tail(96)["date"],
+    y_pred[-96:],
+    "x-",
+    label="DatetimeEncoder() + HGBR prediction",
+)
+ax.plot(
+    X_test.tail(96)["date"],
+    y_pred_wd[-96:],
+    "x-",
+    label="DatetimeEncoder(add_weekday=True) + HGBR prediction",
 )
 
-all_cities = ["Paris", "London"]
-fig, axes = plt.subplots(nrows=len(all_cities), ncols=1, figsize=(12, 9))
-for ax, city in zip(axes, all_cities):
-    mask_prediction = (X_test["city"] == city) & mask_zoom_prediction
-    date_prediction = X_test.loc[mask_prediction]["date.utc"]
-    y_prediction = y_pred[mask_prediction]
-
-    mask_reference = (X["city"] == city) & mask_zoom_reference
-    date_reference = X.loc[mask_reference]["date.utc"]
-    y_reference = y[mask_reference]
-
-    ax.plot(date_reference, y_reference, label="Actual")
-    ax.plot(date_prediction, y_prediction, label="Predicted")
-
-    ax.set(
-        ylabel="NO2",
-        title=city,
-    )
-    ax.legend()
-
+_ = ax.legend()
 plt.show()
+###############################################################################
+# As we can see, adding the weekday yields better predictions on our test set.
 
 
 ###############################################################################
@@ -247,7 +213,7 @@ plt.show()
 # -------------------
 #
 # Using the |DatetimeEncoder| allows us to better understand how the date
-# impacts the NO2 concentration. To this aim, we can compute the
+# impacts the bike sharing demand. To this aim, we can compute the
 # importance of the features created by the |DatetimeEncoder|, using the
 # :func:`~sklearn.inspection.permutation_importance` function, which
 # basically shuffles a feature and sees how the model changes its prediction.
@@ -255,19 +221,17 @@ plt.show()
 ###############################################################################
 from sklearn.inspection import permutation_importance
 
-table_vec = TableVectorizer(datetime=DatetimeEncoder(add_weekday=True))
-
 # In this case, we don't use a pipeline, because we want to compute the
 # importance of the features created by the DatetimeEncoder
-X_transform = table_vec.fit_transform(X)
-feature_names = table_vec.get_feature_names_out()
+X_test_transform = pipeline[:-1].transform(X_test)
 
-model = HistGradientBoostingRegressor().fit(X_transform, y)
-result = permutation_importance(model, X_transform, y, n_repeats=10, random_state=0)
+result = permutation_importance(
+    pipeline[-1], X_test_transform, y_test, n_repeats=10, random_state=0
+)
 
 result = pd.DataFrame(
     dict(
-        feature_names=feature_names,
+        feature_names=X_test_transform.columns,
         std=result.importances_std,
         importances=result.importances_mean,
     )
@@ -280,8 +244,8 @@ plt.tight_layout()
 plt.show()
 
 # %%
-# We can see that the total seconds since Epoch and the hour of the day
-# are the most important feature, which seems reasonable.
+# We can see that the hour of the day, the temperature and the humidity
+# are the most important features, which seems reasonable.
 #
 # Conclusion
 # ----------